Move almost all classes out
