+++ SDLWrapper/ColorMode.cpp
@@ -1,0 +1,14 @@
+#include "ColorMode.h"
+#include "begin_code.h"
+ColorMode::ColorMode(int R, int G, int B)
+{
+    r = R;
+    g = G;
+    b = B;
+}
+
+ColorMode::ColorMode()
+{
+    r = g = b = 0;
+}
+#include "end_code.h"
