+++ SDLWrapper/Cursor.cpp
@@ -1,4 +1,5 @@
 #include "Cursor.h"
+#include "_caster.h"
 #include "begin_code.h"
 //private
 void Cursor::_set(SDL_Cursor* p)
