+++ SDLWrapper/Noncopyable.h
@@ -1,0 +1,9 @@
+#pragma once
+class NonCopyable
+{
+protected:
+    NonCopyable() = default;
+    ~NonCopyable() = default;
+    NonCopyable(const NonCopyable&) = delete;
+    NonCopyable& operator = (const NonCopyable&) = delete;
+};
