+++ SDLWrapper/begin_code.h
@@ -1,5 +1,2 @@
 namespace MiniEngine
 {
-
-inline namespace _SDLWrapper
-{
