+++ SDLWrapper/end_code.h
@@ -1,3 +1,1 @@
-} /// End of namespace MiniEngine::_SDLWrapper
-
 } /// End of namespace MiniEngine
